Bring in latest Konflux default pipeline changes
When we generate a Konflux pipeline in release branch we get the
latest greatest pipeline definition. This PR brings our main branch
pipeline definition in sync with that.

See also #1944 where our EC specific main branch customizations are
applied to a fresh default Konflux pipeline definitions in release
branch.

Notable changes:
- The "build-image-index" task was added
- The "push-dockerfile-oci-ta" task was added
- The netrc workspace was added
- Some obsolete java params were removed

Some of the changes here are inconsequential, e.g. reordering of
params, but they do bring the pipeline closer to indentical with the
default, so future comparisons will be easier.

Handy vimdiff commands for this PR:
  vimdiff +'set ft=yaml' <(git show release-v0.5:.tekton/cli-v05-push.yaml) .tekton/cli-main-ci-push.yaml
  vimdiff +'set ft=yaml' <(git show release-v0.5:.tekton/cli-v05-pull-request.yaml) .tekton/cli-main-ci-pull-request.yaml
  vimdiff +'set ft=yaml' <(git show release-v0.5:.tekton/cli-build.yaml) .tekton/cli-build.yaml

Ref: https://issues.redhat.com/browse/EC-819