Use leaf of package name
The idea is that, for the most part, we only care about the last segment
of a package name. For example, if the rego file defines:

```rego
package policy.release.foo
```

`foo` is the only meaningful name when it comes to inclusion and
exclusions. `policy.release` is ignored in that context. As such, we
should strip that prefix everywhere to avoid confusion.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>