Remove GC before capturing memory profile
The call to runtime.GC forces the Go compiler to run. Making this call
during `OnExit` of the root command means that pretty much all the
memory allocation used during the validate image command, for example,
is erased. This makes the memory profile less interesting when trying to
understand the memory usage of the command.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>