Add Konflux release Pipelines
This commit adds two Tekton Pipelines. One is meant to be used to
release the CLI image via Konflux, the other for releasing the Tekton
Task bundle.

Ideally, a single Pipeline would be used. However, due to limitations in
Konflux release, this would force us to use the same EC policy config
for both images which is currently not supported by EC.

The Pipelines are based on the push-to-external-registry Pipeline
provided by Konflux release (link in code). Some of the Tasks require
changes for a couple of reasons:

1. Some Tasks require ReleasePlanAdmission and ReleaseConfig resource
   references. Neither one of those are used/available during tenant
   releases.

2. Some Tasks assume the image is an OCI image (good thing). However,
   the `tekton bundle push` command creates an Image Manifest with the
   docker media type (kind of deprecated).

To make this saner to maintain, `kustomize` is used to compose the
release Pipelines. `make generate-pipelines` is added to facilitate
running the command.

Additionally, because two EC policies are used, the `.ec` directory
didn't make sense anymore. As such, a new top-level `policies` directory
is used.

Ref: EC-910

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>