Use given source URLs as keys in download cache
We ran into an issue when we started mixing given and pinned source URLs
in the download cache. This has happened because we cached the given
URLs first, mutated the policy to contain the pinned URLs and later
tried to download through the cache via pinned URLs.

This changes so we only use the given source URLs in the download cache
by allowing the source URLs to mutate the policy URLs after they are
downloaded. That is, for each source URL in the policy we have two
states: one containing the given value and used in the download cache,
and second after the download with the resulting resolved/pinned value.

There are two repercussions of this: any resolving/pinning of URLs is
transparent to the callers if they use `source.PolicySourcesFrom`
followed by `url.GetPolicy` -- this allows any use of `Policy` to be
automatically resolved/pinned (no need for Policy pre-processing); and
secondly the Policy is mutated while in execution via `url.GetPolicy`
and careful consideration, hopefully/effectively shielded by using
`source.PolicySourcesFrom`, of the slice mutation -- making sure that
the mutation doesn't happen in the temporary value of the iteration is
required.

Resolves: https://issues.redhat.com/browse/EC-963