update-rpm-lock: use volumes instead of mounts
On systems where SELinux is enabled, it is not possible to modify file
on the host's filesystem unless explicitly enabled by SELinux. The
`--mount` flag does not allow specifying SELinux labels. This commit
changes the podman run command to use `--volume` instead so the labels
can be set as expected.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>