Add script to inspect clair reports
I figured this could be useful for user support to show details
about the CVE detected by clair.

IIUC you can see this kind of information in the clair scan taskrun
logs, but this works if you don't have access to the taskrun logs,
e.g. if they were garbage collected or if you don't have the right
permissions in Konflux.