Build bundle together with CLI
This commit changes the build process so the Tekton bundle is built in
the same build pipeline that builds the CLI image. This has the main
advantage of creating a close relationship between the two images. The
bundle image, for example, references the CLI image by digest. It also
simplifies the build process.

As such, the build pipeline for the bundle is removed.

Ref: EC-911

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>