Add workers flag to `validate input` command
This commit reworks the concurrency model of `validate input` to be like
that found in `validate image`. This commit adds a `--workers` flag,
with a default of 5 that can be used to configure how many simultaneous
fetches occur. This avoids situations where validating multiple input
files can trigger HTTP 429 errors.

Signed-off-by: robnester-rh <rnester@redhat.com>