task: Use arbitrary long timeout instead of zero
The commentary explains the reasoning.

It's not expected to be an issue for Konflux, but theoretically
RHTAP users might have an older EC base image and could some day
migrate to this task definition which would break the task.

Decided to keep the commit separate rather than squash so we can
easily revert it some day.

Ref: https://issues.redhat.com/browse/EC-1030