Update CLI release policy
This commit updates the EC policy config for the CLI image in two ways.

First, the policy config now enforces hermetic builds since the builds
are indeed hermetic 🎉.

Second the policy config no longer enforces the schedule restriction.
Those policy rules are usually meant to avoid releases by a vendor on
restricted dates, e.g. weekends. Since this is being used for a
community release, this is not a concern.

Ref: EC-911

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>