Use consistent rego syntax in prep for opa 1.0
The idea here is that once we update to opa 1.0 we'll need to always
use the new `deny contains foo if` syntax. This change is supposed
to make it easier by using that syntax consistently now.

It requires importing `rego.v1` everywhere, but we do that
consistently for all the rego in ec-policies, so let's also do it
here also.

Note that there were many snapshot changes needed to get the tests
passing, since changes to the rego snippet produce line number
changes in various places, and git sha changes when the rego files
are used with git.

See also #2274 which I'd like to rebase on this once it's merged.

Ref: https://issues.redhat.com/browse/EC-1130