Allow using rekor public key in ec.sigstore func
This commit extends the `ec.sigstore.verify_image` and
`ec.sigstore.verify_attestation` to accept a new option,
`rekor_public_key`. When provided, an offline Rekor verification is
performed on the signature.

The structure of the `opts` parameter for both of those functions also
changes. Now, only the `ignore_rekor` parameter is required. All other
parameters, of type string, are dynamic and no longer required. This
allows the `opts` parameter to omit unnecessary attributes.

Ref: EC-1200

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>