Stop shortening package path codes
This removes some legacy code that would shorten the package path
and use it for the "code".

For example, a deny with short code metadata "bar" in package called
"policy.release.foo", the "code" was shortened down to "foo.bar" for
basically cosmetic reason.

We stopped using those long package names in the rego, so the
shortening no longer does anything, so let's remove it. If a package
did have a long name, it would be included in full in the "code",
which is a better idea anyhow.

Some of the changes to the test rego are not strictly needed, but I
figured it's better to make the test code resemble reality, rather
than have it reflect the obsolete conventions.

Ref: https://issues.redhat.com/browse/EC-945