Use newer wiremock image in generate script
I'm not sure if this change is strictly needed to get things
working, but I was trying a few things while debugging
generate-test-signed-images.sh, and it seems like a good idea to
keep this change either way.

This does move from 2.x to 3.x, so hopefully we're not impacted by
breaking changes. (I considered using 2.35.1-1 but it's looks like
it's a couple of years old, so let's be bood and upgrade.)

See the available tags here: https://hub.docker.com/r/wiremock/wiremock

Ref: https://issues.redhat.com/browse/EC-1247