Use port 5001 not 5000 for registry.local
Not quite sure why this was needed, but it seems to have fixed my
problem. Perhaps it's related to changes in the sigstore/scaffolding
scripts. (See grandparent of this commit.)

`docker ps` showed the port mapping like this for registry.local:
  5000/tcp, 127.0.0.1:5001->5001/tcp

And I was seeing these kind of errors running the script:
  connection refused GET http://registry.local:5000/v2/

Ref: https://issues.redhat.com/browse/EC-1247