Adjust documentationUrl in internal/opa/rule
There was some logic here that depended on the old convention of
long namespaced package names, and hence wasn't working.

Unfortunately this makes the method is even less functional that it
already was, since the subdirectory (i.e. "release", "pipeline",
etc) part is no longer accessible.

As mentioned in the comments I decided to keep the function, but I
added some comments to warn about how broken it is.

Ref: https://issues.redhat.com/browse/EC-945