Properly handle error output when validating image
The 'ec validate image' command now handles error cases more organically, by
collecting them and returning them as a list at the end of the
execution, returning with exit status 1.
The previous implementation caused the command to exit with status 2 and
to output the whole exception in the stderr channel.

Ref: https://issues.redhat.com/browse/EC-1020