Update ubi minimal & switch to latest tag
The reasoning behind switching to the :latest tag, is that
ubi-minimal is single stream, so as soon as :9.6 is pushed, there
are no more updates to :9.5. Using :latest means that we'll
(hopefully) continue to get automated digest bump PRs from
MintMaker/Renovate.

In the unlikely event that something breaks, we'll expect our CI to
catch it.

Digest copy/pasted from:
  https://catalog.redhat.com/software/containers/ubi9/ubi-minimal/615bd9b4075b022acc111bf5?container-tabs=gti

Note that we need it to be the image index digest, not the digest of the
individual image. This should match the digest:
  skopeo inspect --raw --no-tags docker://registry.access.redhat.com/ubi9/ubi-minimal:latest | sha256sum

I ran `hack/update-rpm-lock.sh` but it produced no updates to the
lock file, (unlike the v0.5 version, since we install git-core and
its dependencies there but not there).

Because this brings in the latest rpms (for libtasn1, gnutls and
krb5-libs) in the base image it should remove the vulnerabilities
listed in...

Ref: https://issues.redhat.com/browse/EC-1282