Update ubi minimal & switch to latest tag
The reasoning behind switching to the :latest tag, is that
ubi-minimal is single stream, so as soon as :9.6 is pushed, there
are no more updates to :9.5. Using :latest means that we'll
(hopefully) continue to get automated digest bump PRs from
MintMaker/Renovate.

In the unlikely event that something breaks, we'll expect our CI to
catch it.

Digest copy/pasted from:
  https://catalog.redhat.com/software/containers/ubi9/ubi-minimal/615bd9b4075b022acc111bf5?container-tabs=gti

Note that we need it to be the image index digest, not the digest of the
individual image. This should match the digest:
  skopeo inspect --raw --no-tags docker://registry.access.redhat.com/ubi9/ubi-minimal:latest | sha256sum

Also, update the rpm lock file so we use the newer rpms when
prefetch and hermetically building. Done by running
`hack/update-rpm-lock.sh`

Because this brings in the latest rpms, both in the base image and
in the lock file, it should resolve...

Ref: https://issues.redhat.com/browse/EC-1282