Use standard ubi9/go-toolset
Now that the version we want is available, let's use it. See also
go.mod, which contains `go 1.23.6`.

I decided to do this while looking at previous commit (in
release-v0.5 branch).

It might need to be added to the v0.6 branch and the main branch. I
will have a look shortly.

Not really related to, but done while working on...

Ref: https://issues.redhat.com/browse/EC-1282