feat: improve performance for verifying images
We have been trying to speed up policy evaluation, especially for large
sets of images. As expected, when we increase the number of workers, the
memory requirements also increase to prevent hitting OOM kills on the
pods. In our analysis, however, we saw that memory consumption peaked at
around 6GB even for large sets of images (>35).

While very small sets of images will likely not see a benefit from the
increased worker count, most of the verifications we have seen are
larger sets.

It is a valid option to override these options when consuming this task,
but it would have to be modified for many consumers. I am proposing to
update the defaults as we can quickly see benefits in the evaluation
times, even for modestly large sets of images.

Signed-off-by: arewm <arewm@users.noreply.github.com>

Signed-off-by: arewm <arewm@users.noreply.github.com>