Remove the WORKERS parameter
This task is being used by the release service which overrides the
WORKERS parameter. While this change can be made there, it will take
time for the parameter modification to make it to production.

In order to unblock some users with large Snapshots, I am proposing to
remove the configuration altogether. We already increased the default
resource requirements for the task, so this worker count should suffice.
If this number of workers is too many in the future, users can just
override the resources to further increase them.

Once the release pipeline users can customize the worker count in the
RPA (https://issues.redhat.com/browse/KONFLUX-7850), we can re-enable
this parameter.

Signed-off-by: arewm <arewm@users.noreply.github.com>